feat: add GitHub Actions workflow to fetch data from Notion
This commit adds a new GitHub Actions workflow file, `update.yml`, which is triggered manually. The workflow is responsible for fetching data from Notion and updating the repository with the latest data.

The workflow uses the `ubuntu-latest` runner and consists of the following steps:
- Checkout the repository
- Install Node.js dependencies
- Run a script (`update.cjs`) to fetch data from Notion using the Notion API
- Add the fetched data (`competences.json` and `projects.json`) to the repository
- Commit the changes with the message "Ajout des données depuis Notion"
- Push the changes to the remote repository

Additionally, this commit also includes the following changes:
- Added `@notionhq/client` and `dotenv` as dependencies in `package.json`
- Added `competences.json` and `projects.json` files, which are initially empty
- Added `update.cjs` script, which fetches data from Notion and writes it to the respective JSON files